+++ custom_components/lg_soundbar_plus/__init__.py
@@ -12,6 +12,7 @@ from .coordinator import LGSoundbarCoordinator
 PLATFORMS: list[Platform] = [
     Platform.MEDIA_PLAYER,
     Platform.NUMBER,
+    Platform.SELECT,
     Platform.SWITCH,
 ]